fix: font error message
when building and in the development, it looks fine but a font error occurs.
As I researched, some fonts have the same problem.
This is used to avoid the message problem, since fontFallback is disabled, it may not be considered a stable code.
I will catch up the issue later and for now, to fix, apply the options to the font.