refactor: update db connection logic
executeDB calls the callback function that it receives as its parameter then close the connection.

By accessing db through this function, it ensures a connection is closed after using the connection.

Addtionally, you can use the `cache` option if you want. But the cache has benefits after the first call.