refactor: connection cache logic
global.mongoConn is only used for cache purpose and once it's made, persist the connection regardless of the useCache option.

In the previous logic, the cached connection is cleared depending on the useCache option.

Added events to make sure that the connection is continuously connected, but as this logic isn't tested, whether the events keep the connection is not sure.